+++ apps/amiapp_flutter/lib/src/screens/dashboard.dart
@@ -303,6 +303,8 @@ enum _ActionItem {
   deviceAttributes,
   profileAttributes,
   inlineMessages,
+  inboxMessages,
+  location,
   showPushPrompt,
   showLocalPush,
   signOut,
@@ -321,6 +323,10 @@ extension _ActionNames on _ActionItem {
         return 'Set Profile Attribute';
       case _ActionItem.inlineMessages:
         return 'Test Inline Messages';
+      case _ActionItem.inboxMessages:
+        return 'Inbox Messages';
+      case _ActionItem.location:
+        return 'Test Location';
       case _ActionItem.showPushPrompt:
         return 'Show Push Prompt';
       case _ActionItem.showLocalPush:
@@ -342,6 +348,10 @@ extension _ActionNames on _ActionItem {
         return 'Profile Attribute Button';
       case _ActionItem.inlineMessages:
         return 'Inline Messages Button';
+      case _ActionItem.inboxMessages:
+        return 'Inbox Messages Button';
+      case _ActionItem.location:
+        return 'Location Button';
       case _ActionItem.showPushPrompt:
         return 'Show Push Prompt Button';
       case _ActionItem.showLocalPush:
@@ -363,6 +373,10 @@ extension _ActionNames on _ActionItem {
         return Screen.profileAttributes;
       case _ActionItem.inlineMessages:
         return Screen.inlineMessages;
+      case _ActionItem.inboxMessages:
+        return Screen.inboxMessages;
+      case _ActionItem.location:
+        return Screen.locationTest;
       case _ActionItem.showPushPrompt:
         return null;
       case _ActionItem.showLocalPush: