+++ lib/customer_io_config.dart
@@ -1,3 +1,4 @@
 export 'config/customer_io_config.dart';
 export 'config/in_app_config.dart';
+export 'config/location_config.dart';
 export 'config/push_config.dart';