+++ lib/location/_native_constants.dart
@@ -1,0 +1,11 @@
+/// Methods specific to Location module.
+class NativeMethods {
+  static const String setLastKnownLocation = "setLastKnownLocation";
+  static const String requestLocationUpdate = "requestLocationUpdate";
+}
+
+/// Method parameters specific to Location module.
+class NativeMethodParams {
+  static const String latitude = "latitude";
+  static const String longitude = "longitude";
+}